+++ src/tracker/controller/HistoryManager.java
@@ -9,7 +9,7 @@ public interface HistoryManager {
     // метод добавляет просмотренную задачу в список
     void add(Task task);
     // метод удаляет просмотр из истории
-    void remove(long id);
+    void remove(int id);
     // метод возвращает список просмотренных задач
     List<Task> getHistory();
 }
